chore: move htmlToMarkdown to scripts/utils/migrate folder
